VTCCA-904 change organisation id to long (#290)
VTCCA-904 Make organisationId/personId a Long

+++ app/models/IndividualAccount.scala
@@ -25,7 +25,7 @@ object IndividualAccountSubmission {
 
 case class IndividualAccount(externalId: String, trustId: String, organisationId: Long, details: IndividualDetails)
 
-case class DetailedIndividualAccount(externalId: String, trustId: String, organisationId: Long, individualId: Int, details: IndividualDetails) {
+case class DetailedIndividualAccount(externalId: String, trustId: String, organisationId: Long, individualId: Long, details: IndividualDetails) {
   def toIndividualAccount: IndividualAccount = IndividualAccount(externalId, trustId, organisationId, details)
 }
